Still More Comment Blocking
Getting a clearer idea of what needs to happen. Still, some mysteries
yet remain, particularly in some of the structure of the EBTB.

Also updating alpha.h here (just adding a comment, but that should make
it easier to identify later. We may need to come back to it for more
information on the ISA.)